[Bug 156418799]Add basic endpoints
