+++ tests.py
@@ -4,6 +4,7 @@ from run import app
 from flask import jsonify
 from api.models import Book, User
 
+
 class TestBooksModel(unittest.TestCase):
 
     def setUp(self):
@@ -11,7 +12,8 @@ class TestBooksModel(unittest.TestCase):
 
         :return:
         """
-        self.book = Book('Game of Thrones', 'George R.R. Martin', '3878749').createbook()
+        self.book = Book('Game of Thrones',
+                         'George R.R. Martin', '3878749').createbook()
 
     def tearDown(self):
         """
@@ -26,7 +28,8 @@ class TestBooksModel(unittest.TestCase):
         :return:
         """
 
-        res = Book('Game of Thrones', 'George R.R. Martin', '363837').createbook()
+        res = Book('Game of Thrones', 'George R.R. Martin',
+                   '363837').createbook()
         self.assertEqual(res, {'Success': 'Book Created Successfully'})
 
     def test_book_already_exists(self):
@@ -34,7 +37,8 @@ class TestBooksModel(unittest.TestCase):
 
         :return:
         """
-        res = Book('Game of Thrones', 'George R.R. Martin', '3878749').createbook()
+        res = Book('Game of Thrones', 'George R.R. Martin',
+                   '3878749').createbook()
         self.assertEqual(res, {'Message': 'Book Already Exists'})
 
     def test_book_does_not_exist(self):
@@ -50,7 +54,8 @@ class TestBooksModel(unittest.TestCase):
 
         :return:
         """
-        res = Book.updatebook('1', {'title': 'GOT', 'author': 'George', 'isbn': '827890'})
+        res = Book.updatebook(
+            '1', {'title': 'GOT', 'author': 'George', 'isbn': '827890'})
         self.assertEqual(res, {'Message': 'Book Does Not Exist'})
 
     # def test_book_delete_successful(self):
@@ -80,7 +85,8 @@ class TestBookAPI(unittest.TestCase):
 
         app.testing = True
         self.app = app.test_client()
-        self.book = Book(title="Book Title", author="Book Author", isbn="456788")
+        self.book = Book(title="Book Title",
+                         author="Book Author", isbn="456788")
 
     def tearDown(self):
         """
@@ -132,7 +138,7 @@ class TestBookAPI(unittest.TestCase):
         Book(title="Book Title", author="Book Author", isbn="456788")
         res = self.app.get('/api/v1/books/1>')
         print(res.data)
-        self.assertEqual(res.status_code, 404)#Revisit
+        self.assertEqual(res.status_code, 404)  # Revisit
 
     def test_get_book_by_id_fail(self):
         """
@@ -148,9 +154,9 @@ class TestBookAPI(unittest.TestCase):
         :return:
         """
         payload = {
-            "title":"New Title",
-            "author":"New Author",
-            "isbn":"3786376"
+            "title": "New Title",
+            "author": "New Author",
+            "isbn": "3786376"
         }
         res = self.app.put('/api/v1/books/1>', data=json.dumps(payload))
         self.assertEqual(res.status_code, 200)
@@ -161,7 +167,7 @@ class TestBookAPI(unittest.TestCase):
         :return:
         """
         payload = {
-            
+
         }
         res = self.app.put('/api/v1/books/1>', data=json.dumps(payload))
         self.assertEqual(res.status_code, 400)
@@ -248,9 +254,76 @@ class TestUserModel(unittest.TestCase):
 
         :return:
         """
-        book = Book(title='Book Title', author='Book Author', isbn="64368").createbook()
+        book = Book(title='Book Title', author='Book Author',
+                    isbn="64368").createbook()
         res = User.borrowBook("10")
         self.assertEqual(res, {'Message': 'Book Does Not Exist'})
+
+
+class TestUserAPI(unittest.TestCase):
+
+    def setUp(self):
+        """
+
+        :return:
+        """
+        app.testing = True
+        self.app = app.test_client()
+        self.user = User('dmwangi', 'password123', 'True').createUser()
+
+    def tearDown(self):
+        """
+
+        :return:
+        """
+        app.testing = False
+        self.app = None
+
+    def test_register_user_bad_request(self):
+        """
+        Tests whether the register user API endpoint can pass a Bad Request(Missing User Information)
+        Send a post request to register user API with no user information(Bad Request).
+        :return: 400 Bad Request Status Response
+        """
+        res = self.app.post('/api/v1/auth/register', data={})
+        self.assertEqual(res.status_code, 400)
+
+    def test_register_user_success(self):
+        """
+        Tests 201 status code response when user has been created successfully
+        :return: 201 Created Status Response
+        """
+        payload = {
+            "username": "username",
+            "password": "password",
+            "admin": "True"
+        }
+        res = self.app.post('/api/v1/auth/register', data=json.dumps(payload))
+        self.assertEqual(res.status_code, 201)
+
+    def test_get_all_user_api_exits(self):
+        """
+
+        :return:
+        """
+        res = self.app.get('/api/v1/users')
+        self.assertEqual(res.status_code, 200)
+
+    def test_borrow_book_api_endpoint(self):
+        """
+
+        :return:
+        """
+        res = self.app.post('/api/v1/users/books/<string:book_id>')
+        self.assertEqual(res.status_code, 200)
+
+    def test_update_password_exists(self):
+        """
+
+        :return:
+        """
+        res = self.app.post('/api/v1/auth/reset-password/<string:user_id>')
+        self.assertEqual(res.status_code, 200)
 
 
 if __name__ == '__main__':
